pyproject.toml
[project]
name = "pointint"
version = "0.1.0"
requires-python = ">=3.9,<3.13"
dependencies = [
    "torch==2.8.0+cu126",
    "torchvision==0.23.0+cu126",
    "kaolin==0.18.0",
    "warp-lang==1.9.0",
    "loguru>=0.7.2",
    "tqdm>=4.66.5",
    "pip>=24.0",
    "pyside6",
    "pyvista",
    "pyvistaqt",
    "viser>=0.2.0",
]

[project.optional-dependencies]
# Minimal dependencies for intersection module (works on macOS without kaolin)
intersection = [
    "warp-lang==1.9.0",
    "numpy>=1.24.0",
    "pytest>=7.0.0",
    "scipy>=1.10.0",
]

[dependency-groups]
# Dependency group for intersection module testing (no kaolin needed)
intersection-test = [
    "warp-lang==1.9.0",
    "numpy>=1.24.0",
    "pytest>=7.0.0",
    "scipy>=1.10.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["pointint"]

[tool.setuptools.package-data]
pointint = ["py.typed"]

[tool.uv]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
kaolin = { index = "kaolin-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "kaolin-cu126"
url = "https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.8.0_cu126.html"
format = "flat"
